+++ CHANGELOG.md
@@ -59,6 +59,7 @@ Alternatively view it [online on GitHub](https://github.com/cumulusmx/CumulusMX/
 - Fixed the web tag `ProgramUpTimeMs`, it now returns a integer value as originally intended
 - Dashboard Select-a-Period AQ PM2.5 chart only showed data from the last N days - N defined by the recent data setting
 - MQTT IP protocol version not saving to Cumulus.ini
+- MQTT Protocol version - removed Auto detect as it does not work
 
 
 ### Package Updates